changed some directory names and added handlebars
