+++ .env.example
@@ -1,0 +1,3 @@
+DB_NAME='games_db'
+DB_USER=''
+DB_PASSWORD=''
