+++ db/schema.sql
@@ -1,0 +1,6 @@
+-- DROP DATABASE
+DROP DATABASE IF EXISTS games_db;
+
+-- CREATE DATABASE
+CREATE DATABASE games_db;
+
