+++ package.json
@@ -1,0 +1,21 @@
+{
+  "name": "project-2-video-game-reviews",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server handlebars.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "bcrypt": "^5.1.1",
+    "connect-session-sequelize": "^7.1.7",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-handlebars": "^7.1.2",
+    "mysql2": "^3.6.3",
+    "sequelize": "^6.34.0"
+  }
+}
