+++ public/css/style.css
@@ -1,0 +1,56 @@
+header,
+main, 
+footer {
+  border: 2px solid red;
+  padding: 10px;
+}
+
+h1, h2 {
+  border: 2px solid orange;
+  padding: 10px;
+}
+
+colors {
+  color: #8921C2;
+  color: #fe39a4;
+  color: #fffdbb;
+  color: #53e8d4;
+  color: #25c4f8;
+
+}
+
+.neonText {
+  color: #fff;
+  text-shadow:
+    /* White glow */
+    0 0 7px #fff,
+    0 0 10px #fff,
+    0 0 21px #fff,
+    /* Green glow */
+    0 0 42px #0fa,
+    0 0 82px #0fa,
+    0 0 92px #0fa,
+    0 0 102px #0fa,
+    0 0 151px #0fa;
+}
+
+@keyframes flicker {
+  0%, 18%, 22%, 25%, 53%, 57%, 100% {
+    text-shadow:
+      0 0 4px #fff,
+      0 0 11px #fff,
+      0 0 19px #fff,
+      0 0 40px #0fa,
+      0 0 80px #0fa,
+      0 0 90px #0fa,
+      0 0 100px #0fa,
+      0 0 150px #0fa;
+  }
+  20%, 24%, 55% {       
+    text-shadow: none;
+  }
+}
+
+h1 {
+  animation: flicker 1.5s infinite alternate;     
+}
