+++ public/index.html.example
@@ -1,0 +1,46 @@
+<!-- 
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Document</title>
+  link to bootstrap
+  <link
+    rel="stylesheet"
+    href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+  />
+  font awesome
+  <link
+    rel="stylesheet"
+    href="https://use.fontawesome.com/releases/v5.8.1/css/all.css"
+    integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf"
+    crossorigin="anonymous"
+  />
+  <link rel="stylesheet" href="./css/style.css">
+</head>
+
+  <body>
+    <header>
+      <h1>Game Reviewer</h1>
+    </header>
+    <main>
+      <div>
+        <h2>SECTION 1</h2>
+      </div>
+      <div>
+        <h2>SECTION 2</h2>
+      </div>
+      <div>
+        <h2>SECTION 3</h2>
+      </div>
+    </main>
+    <footer>
+      &copy; Copyright 2023 Group 1 
+    </footer>
+    JQUERY
+    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
+  </body>
+
+</html> -->
